PST: Updated SDK to recent SVN version.

+++ src/modules/pst/pstsdk/ndb/node.h
@@ -13,7 +13,6 @@
 #include <algorithm>
 #include <memory>
 #include <cassert>
-#include <functional>
 #include <boost/iterator/transform_iterator.hpp>
 #include <boost/iostreams/concepts.hpp>
 #ifdef _MSC_VER
@@ -821,7 +820,7 @@ public:
         { return m_subnode_info[pos].first; }
     subnode_block* get_child(uint pos);
     const subnode_block* get_child(uint pos) const;
-    uint num_values() const { return (uint) m_subnode_info.size(); }
+    uint num_values() const { return m_subnode_info.size(); }
     
 private:
     std::vector<std::pair<node_id, block_id> > m_subnode_info;           //!< Info about the sub-blocks
@@ -859,7 +858,7 @@ public:
     const node_id& get_key(uint pos) const
         { return m_subnodes[pos].first; }
     uint num_values() const
-        { return (uint) m_subnodes.size(); }
+        { return m_subnodes.size(); }
 
 private:
     std::vector<std::pair<node_id, subnode_info> > m_subnodes;   //!< The actual subnode information
